Split up C and C++ header files by types
