+++ example/c/locale_ffi_ICU4XLocale.h
@@ -1,0 +1,21 @@
+#ifndef locale_ffi_ICU4XLocale_H
+#define locale_ffi_ICU4XLocale_H
+#include <stdio.h>
+#include <stdint.h>
+#include <stddef.h>
+#include <stdbool.h>
+#include "diplomat_runtime.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+typedef struct ICU4XLocale ICU4XLocale;
+
+ICU4XLocale* ICU4XLocale_new(const char* name_data, size_t name_len);
+void ICU4XLocale_destroy(ICU4XLocale* self);
+#ifdef __cplusplus
+}
+#endif
+#endif
