+++ example/cpp/data_provider_ffi_ICU4XDataProvider.hpp
@@ -1,0 +1,36 @@
+#ifndef data_provider_ffi_ICU4XDataProvider_HPP
+#define data_provider_ffi_ICU4XDataProvider_HPP
+#include <stdint.h>
+#include <stddef.h>
+#include <stdbool.h>
+#include <algorithm>
+#include <memory>
+#include <optional>
+#include <variant>
+#include "diplomat_runtime.hpp"
+
+namespace capi {
+#include "data_provider_ffi_ICU4XDataProvider.h"
+}
+
+class ICU4XDataProvider;
+
+struct ICU4XDataProviderDeleter {
+  void operator()(capi::ICU4XDataProvider* l) const noexcept {
+    capi::ICU4XDataProvider_destroy(l);
+  }
+};
+class ICU4XDataProvider {
+ public:
+  static ICU4XDataProvider new_static();
+  inline const capi::ICU4XDataProvider* AsFFI() const { return this->inner.get(); }
+  inline capi::ICU4XDataProvider* AsFFIMut() { return this->inner.get(); }
+  ICU4XDataProvider(capi::ICU4XDataProvider* i) : inner(i) {}
+ private:
+  std::unique_ptr<capi::ICU4XDataProvider, ICU4XDataProviderDeleter> inner;
+};
+
+ICU4XDataProvider ICU4XDataProvider::new_static() {
+  return ICU4XDataProvider(capi::ICU4XDataProvider_new_static());
+}
+#endif
