+++ example/cpp/decimal_ffi_ICU4XFixedDecimalFormat.h
@@ -1,0 +1,28 @@
+#ifndef decimal_ffi_ICU4XFixedDecimalFormat_H
+#define decimal_ffi_ICU4XFixedDecimalFormat_H
+#include <stdio.h>
+#include <stdint.h>
+#include <stddef.h>
+#include <stdbool.h>
+#include "diplomat_runtime.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+typedef struct ICU4XFixedDecimalFormat ICU4XFixedDecimalFormat;
+#include "locale_ffi_ICU4XLocale.h"
+#include "data_provider_ffi_ICU4XDataProvider.h"
+#include "decimal_ffi_ICU4XFixedDecimalFormatOptions.h"
+#include "decimal_ffi_ICU4XFixedDecimalFormatResult.h"
+#include "fixed_decimal_ffi_ICU4XFixedDecimal.h"
+
+ICU4XFixedDecimalFormatResult ICU4XFixedDecimalFormat_try_new(const ICU4XLocale* locale, const ICU4XDataProvider* provider, ICU4XFixedDecimalFormatOptions options);
+
+void ICU4XFixedDecimalFormat_format_write(const ICU4XFixedDecimalFormat* self, const ICU4XFixedDecimal* value, DiplomatWriteable* write);
+void ICU4XFixedDecimalFormat_destroy(ICU4XFixedDecimalFormat* self);
+#ifdef __cplusplus
+}
+#endif
+#endif
